Move Arviss-encoding to dcode/
Why "dcode"?
It's short for deconstructed code.
