+++ include/arviss/dcode/arviss_encoder.h
